pyproject.toml
[project]
name = "pds"
version = "0.1.0"
description = "Please Deploy Stuff - Minimal CLI to provision VMs, run migrations, and deploy apps from a single YAML using OpenTofu/Terraform and Ansible."
readme = "README.md"
authors = [{name = "Mikhail Kraevskii"}]
license = {file = "LICENSE"}
keywords = ["deployment", "devops", "terraform", "ansible", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "platformdirs>=4.0.0",
]

[project.scripts]
pds = "pds.cli:app"

[project.urls]
Homepage = "https://github.com/mikhailkraevskii/pds"
Repository = "https://github.com/mikhailkraevskii/pds"
Issues = "https://github.com/mikhailkraevskii/pds/issues"

[dependency-groups]
dev = [
    "ruff>=0.12.8",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pyright>=1.1.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "safety>=3.0.0",
    "pydocstyle>=6.3.0",
    "nox>=2024.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--cov=pds",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.pydocstyle]
convention = "google"
add_ignore = ["D100", "D104"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "I", "N", "D", "UP", "S"]
ignore = ["D100", "D104", "D203", "D213"]

[tool.ruff.format]
quote-style = "double"            # или "single"
indent-style = "space"            # "space" или "tab"
line-ending = "lf"                 # "lf", "crlf", "native"
docstring-code-format = true       # форматировать код в докстрингах